# Remove branch in optimized is_ascii