@thomcc 2020-07-20
> either way we can get unaligned_tail have a tiny chance of 1/8 on a 64 bit machine. I don't think we should bet on that, the probability is worse than dice.

I think it's the other way around -- the tail being aligned has a probability of 1/8. I only added a bench for one possible misalignment, but there are several.