@pickfire 2020-07-20
> I think it's the other way around -- the tail being aligned has a probability of 1/8. I only added a bench for one possible misalignment, but there are several.

Ah, I meant the tail being aligned have a probability of 1/8. Sorry if I am not being clear.

> Edit: To be clear, I'm not opposed to this change in principal, but I think it's worth keeping in mind that the unaligned tail case is the common case (an argument could be made either way about unaligned head).

I don't understand how unaligned tail with a probability of 1/8 is the common case (> 1/2)? Isn't it the other way around? Most of these checks I believe is either based on user input or have it being read somewhere, I don't think users will think to type in numbers of 8 to be aligned with the compiler.