@thomcc 2020-07-20
> I don't understand how unaligned tail with a probability of 1/8 is the common case (> 1/2)? Isn't it the other way around? Most of these checks I believe is either based on user input or have it being read somewhere, I don't think users will think to type in numbers of 8 to be aligned with the compiler.

That's exactly why unaligned tail is the common case. Strings whose length modulo 8 is 1, 2, 3, 4, 5, 6, or 7 all have unaligned tails.