@pickfire 2020-07-20
> That's exactly why unaligned tail is the common case. Strings whose length modulo 8 is 1, 2, 3, 4, 5, 6, or 7 all have unaligned tails.

Oh, I thought it is the other way around. Sorry my bad. But still, won't we have more consistent worst case?