@nagisa 2020-07-21
> It improves medium and short by 1ns but regresses unaligned_tail by 2ns

This amount of improvement sounds like its most likely within the error margin. At this point we probably would need to start measuring cycles (with e.g. llvm-mca), rather than nanoseconds.

Much like @thomcc I’m worried that that the impact on having an unconditional unaligned load will move the performance hit outside of the error margin on architectures where unaligned loads are expensive (and possibly simulated in generated code).

One thing we could try to mitigate the cost somewhat is replace the unaligned load with a plain byte-by-byte loop if we are making that part of code unconditional... though it would probably make the improvement also vanish...

All that said, I guess I'm not super against the idea of landing this, but it does not look like a slam dunk either.