@thomcc 2020-07-22
> One thing we could try to mitigate the cost somewhat is replace the unaligned load with a plain byte-by-byte loop if we are making that part of code unconditional... though it would probably make the improvement also vanish...

It's not uncommon for these byte-for-byte loops to be slower than the unaligned load on ARM, so I don't know if I'd assume right out of the gate that that's a better option.