@pickfire 2020-07-22
Maybe we do a benchmark on ARM? Should I try it on my raspberry pi and see which is the fastest?