@lzutao 2020-07-22
Gcc compile farm has ARM machines, but guess you are already know that.